Refactor crates
