+++ micro-sync/sync-model/Cargo.toml
@@ -1,9 +1,13 @@
 [package]
-name = "sync-model"
-version = "0.1.0"
 authors = ["Terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
+name = "sync-model"
+version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+core-model = {git = "https://github.com/Terkwood/BUGOUT", branch = "fix/shared-structs"}
+move-model = {git = "https://github.com/Terkwood/BUGOUT", branch = "fix/shared-structs"}
+serde = "1.0.116"
+serde_derive = "1.0.116"
