+++ micro-sync/sync-model/src/lib.rs
@@ -1,7 +1,15 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
+extern crate core_model;
+extern crate move_model;
+extern crate serde_derive;
+
+pub mod api;
+
+use serde_derive::{Deserialize, Serialize};
+
+use move_model::{Coord, Player};
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
+pub struct Move {
+    pub player: Player,
+    pub coord: Option<Coord>,
+    pub turn: u32,
 }
